Move Cert Types to s2n_certificate.h, and add Client Auth params to s2n_config
