+++ crypto/s2n_certificate.h
@@ -19,7 +19,22 @@
 
 #include "crypto/s2n_rsa.h"
 #include "stuffer/s2n_stuffer.h"
-#include "tls/s2n_config.h"
+
+#define S2N_MAX_SERVER_NAME 256
+
+struct s2n_cert_chain {
+    struct s2n_blob cert;
+    struct s2n_cert_chain *next;
+};
+
+struct s2n_cert_chain_and_key {
+    uint32_t chain_size;
+    struct s2n_cert_chain *head;
+    struct s2n_rsa_private_key private_key;
+    struct s2n_blob ocsp_status;
+    struct s2n_blob sct_list;
+    char server_name[S2N_MAX_SERVER_NAME];
+};
 
 /* RFC's that define below values:
  *  - https://tools.ietf.org/html/rfc5246#section-7.4.4
@@ -45,6 +60,21 @@ struct s2n_cert_public_key {
         /* TODO: Support other Public Key Types (Eg ECDSA) */
     } public_key;
 };
+
+typedef enum { S2N_CERT_AUTH_NONE, S2N_CERT_AUTH_REQUIRED } s2n_cert_auth_type;
+
+/* Verifies the Certificate Chain of trust and places the leaf Certificate's Public Key in the public_key_out parameter.
+*
+* Does not perform any hostname validation, which is still needed in order to completely validate a Certificate.
+*
+* @param cert_chain_in The DER formatted full chain of certificates recieved
+* @param public_key_out The public key that should be updated with the key extracted from the certificate
+* @param context A pointer to any caller defined context data
+*
+* @return The function should return 0 if Certificate is trusted and public key extraction was successful, and less than
+*         0 if the Certificate is untrusted, or there was some other error.
+*/
+typedef int verify_cert_trust_chain(struct s2n_blob *cert_chain_in, struct s2n_cert_public_key *public_key_out, void *context);
 
 int s2n_send_cert_chain(struct s2n_stuffer *out, struct s2n_cert_chain_and_key *chain);
 int s2n_cert_public_key_set_cert_type(struct s2n_cert_public_key *cert_pub_key, s2n_cert_type cert_type);
