+++ tls/s2n_config.c
@@ -67,6 +67,11 @@ int get_nanoseconds_since_epoch(void *data, uint64_t * nanoseconds)
 
 #endif
 
+int deny_all_certs(struct s2n_blob *x509_der_cert, struct s2n_cert_public_key *public_key, void *context)
+{
+    S2N_ERROR(S2N_ERR_CERT_UNTRUSTED);
+}
+
 struct s2n_config s2n_default_config = {
     .cert_and_key_pairs = NULL,
     .cipher_preferences = &cipher_preferences_20170210,
@@ -96,6 +101,12 @@ struct s2n_config *s2n_config_new(void)
     new_config->cache_delete = NULL;
     new_config->cache_delete_data = NULL;
     new_config->ct_type = S2N_CT_SUPPORT_NONE;
+
+    /* By default, only the client will authenticate the Server's Certificate. The Server does not request or
+     * authenticate any client certificates. */
+    new_config->client_cert_auth_type = S2N_CERT_AUTH_NONE;
+    new_config->verify_cert_chain_cb = deny_all_certs;
+    new_config->verify_cert_context = NULL;
 
     GUARD_PTR(s2n_config_set_cipher_preferences(new_config, "default"));
