+++ tls/s2n_config.h
@@ -15,29 +15,14 @@
 
 #pragma once
 
+#include "crypto/s2n_certificate.h"
 #include "crypto/s2n_rsa.h"
 #include "crypto/s2n_dhe.h"
 
 #include "utils/s2n_blob.h"
 #include "api/s2n.h"
 
-#define S2N_MAX_SERVER_NAME 256
-
 struct s2n_cipher_preferences;
-
-struct s2n_cert_chain {
-    struct s2n_blob cert;
-    struct s2n_cert_chain *next;
-};
-
-struct s2n_cert_chain_and_key {
-    uint32_t chain_size;
-    struct s2n_cert_chain *head;
-    struct s2n_rsa_private_key private_key;
-    struct s2n_blob ocsp_status;
-    struct s2n_blob sct_list;
-    char server_name[S2N_MAX_SERVER_NAME];
-};
 
 struct s2n_config {
     struct s2n_dh_params *dhparams;
@@ -61,6 +46,10 @@ struct s2n_config {
     int (*cache_delete) (void *data, const void *key, uint64_t key_size);
     void *cache_delete_data;
     s2n_ct_support_level ct_type;
+
+    s2n_cert_auth_type client_cert_auth_type;
+    verify_cert_trust_chain *verify_cert_chain_cb;
+    void *verify_cert_context;
 };
 
 extern struct s2n_config s2n_default_config;
