+++ tls/s2n_connection.c
@@ -41,11 +41,6 @@
 #include "utils/s2n_blob.h"
 #include "utils/s2n_mem.h"
 
-int deny_all_certs(struct s2n_blob *x509_der_cert, struct s2n_cert_public_key *public_key, void *context)
-{
-    S2N_ERROR(S2N_ERR_CERT_UNTRUSTED);
-}
-
 /* Accept all RSA Certificates is unsafe and is only used in the s2n Client */
 int accept_all_rsa_certs(struct s2n_blob *cert_chain_in, struct s2n_cert_public_key *public_key_out, void *context)
 {
@@ -97,12 +92,13 @@ struct s2n_connection *s2n_connection_new(s2n_mode mode)
      * which is ok, as blob.data is always aligned.
      */
     conn = (struct s2n_connection *)(void *)blob.data;
+    conn->config = &s2n_default_config;
 
     /* By default, only the client will authenticate the Server's Certificate. The Server does not request or
      * authenticate any client certificates. */
-    conn->client_cert_auth_type = S2N_CERT_AUTH_NONE;
-    conn->verify_cert_chain_cb = deny_all_certs;
-    conn->verify_cert_context = NULL;
+    conn->client_cert_auth_type = conn->config->client_cert_auth_type;
+    conn->verify_cert_chain_cb = conn->config->verify_cert_chain_cb;
+    conn->verify_cert_context = conn->config->verify_cert_context;
 
     if (mode == S2N_CLIENT) {
         /* At present s2n is not suitable for use in client mode, as it
@@ -120,7 +116,6 @@ struct s2n_connection *s2n_connection_new(s2n_mode mode)
 
     conn->mode = mode;
     conn->blinding = S2N_BUILT_IN_BLINDING;
-    conn->config = &s2n_default_config;
     conn->close_notify_queued = 0;
     conn->session_id_len = 0;
     conn->send = NULL;
